Create RRAA.C
The main objective is to develop a new approach for round robin CPU scheduling Algorithm,
which improves the performance of CPU in Real time Operating Systems.
It is designed especially for time sharing systems. The ready queue is treated as a circular
queue.It retains the advantage of round robin in reducing starvation and also integrates the
advantage of priority scheduling.